requirements.txt
python>=3.8
torch>=1.10
torchvision>=0.11
opencv-python
Pillow
numpy
scikit-learn
tqdm
onnx
onnxruntime
matplotlib